requirements.txt
asgiref==3.11.1
Django==5.2.7
django-tailwind==4.4.2
et_xmlfile==2.0.0
google-genai==1.69.0
lxml==6.0.2
openpyxl==3.1.5
pytailwindcss==0.3.0
python-docx==1.2.0
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
